+++ tests/unit/adapters/test_ram_only_extra.py
@@ -1,6 +1,6 @@
 from hyperon_das_atomdb.adapters.ram_only import InMemoryDB
 from tests.unit.fixtures import in_memory_db  # noqa: F401
-from tests.unit.test_database import check_handle
+from tests.unit.test_database_public_methods import check_handle
 
 
 class TestRamOnlyExtra:
